+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [1.5.0] 2021-03-05
+
 - New command **hardis:org:select**
 - New command **hardis:work:resetselection**
 - **hardis:work:save**: Upgrade package.xml and destructiveChanges.xml from git diff